refactor global types name
