+++ src/service/impl/trips/graphql/jp3/trip.graphql-gen.ts
@@ -1,4 +1,4 @@
-import * as Types from '../../../../../graphql/journey-types_v3';
+import * as Types from '../../../../../graphql/journeyplanner-types_v3';
 
 import { DocumentNode } from 'graphql';
 import gql from 'graphql-tag';
